requirements.txt
Django==1.7.1
Markdown==2.5.2
Pillow==2.6.1
django-admin-tools==0.5.2
django-appconf==0.6
django-compress==1.0.1
django-compressor==1.4
django-embed-video==0.11
django-extensions==1.4.8
django-filter==0.9.1
django-generic-images==0.36.1
django-grappelli==2.6.3
django-imagekit==3.2.4
djangorestframework==3.0.0
fudge==1.0.3
image==1.3.4
mysqlclient==1.3.4
pilkit==1.1.12
requests==2.5.0
simplejson==3.6.5
six==1.8.0
soundcloud==0.4.1
